Added import api

+++ api/models/documentary.go
@@ -2,11 +2,11 @@ package models
 
 // Documentary is a data model
 type Documentary struct {
-	Type        int    `json:"type,omitempty"`
+	Type        int64  `json:"type,omitempty"`
 	ImdbID      string `json:"imdb_id,omitempty"`
 	Name        string `json:"name,omitempty"`
-	Year        int    `json:"year,omitempty"`
-	Score       int    `json:"score,omitempty"`
+	Year        int64  `json:"year,omitempty"`
+	Score       int64  `json:"score,omitempty"`
 	Director    string `json:"director,omitempty"`
 	Watched     bool   `json:"watched,omitempty"`
 	DateWatched string `json:"date_watched,omitempty"`
